Moved info and response classes to respective packages, created Validates annotation
