+++ src/main/java/api/utils/validation/Validates.java
@@ -1,0 +1,9 @@
+package api.utils.validation;
+
+/**
+ * Created by coon on 08.03.17.
+ */
+
+public @interface Validates {
+    String name();
+}
